Better networkx integration

+++ test_workforce.py
@@ -1,16 +1,17 @@
 #!/usr/bin/env python3
-import unittest
-import time
-import pandas as pd
+import csv
 import os
+import time
+import unittest
 import workforce
 
 
 class TestWorker(unittest.TestCase):
     def cleanup(self):
-        self.testfile = "test_plan.csv"
-        if os.path.isfile(self.testfile):
-            os.remove(self.testfile)
+        self.planfile = "test_plan.csv"
+        if os.path.isfile(self.planfile): os.remove(self.planfile)
+        self.testfile = "multiproc_test"
+        if os.path.isfile(self.testfile): os.remove(self.testfile)
 
     def setUp(self):
         self.cleanup()
@@ -20,18 +21,17 @@ class TestWorker(unittest.TestCase):
 
     def test_schema(self):
         # Tests a simple plan to make sure multiprocessing works
-        multi_test = "multiproc_test"
-        test_array = pd.DataFrame([
+        test_array = [
             ["echo 1", "echo 2"],
             ["echo 2", "touch multiproc_test"],
-            ["echo 2", "echo 3"]])
-        test_array.to_csv(self.testfile, header=False, index=False)
+            ["echo 2", "echo 3"]]
+        with open(self.testfile, "w") as ff:
+            writer = csv.writer(ff)
+            writer.writerows(test_array)
         worker = workforce.worker(self.testfile)
         worker.run()
         time.sleep(1)
-        assert os.path.exists(multi_test)
-        if os.path.exists(multi_test):
-            os.remove(multi_test)
+        assert os.path.exists(self.testfile)
 
 
 if __name__ == '__main__':
